+++ yellowstone-grpc-tools/src/bin/grpc-kafka.rs
@@ -226,16 +226,13 @@ impl ArgsAction {
         tokio::pin!(kafka_error_rx);
 
         // Create gRPC client & subscribe
-        let mut client = GeyserGrpcClient::connect_with_timeout(
-            config.endpoint,
-            config.x_token,
-            None,
-            Some(Duration::from_secs(10)),
-            Some(Duration::from_secs(5)),
-            false,
-        )
-        .await?;
-        let mut geyser = client.subscribe_once2(config.request.to_proto()).await?;
+        let mut client = GeyserGrpcClient::build_from_shared(config.endpoint)?
+            .x_token(config.x_token)?
+            .connect_timeout(Duration::from_secs(10))
+            .timeout(Duration::from_secs(5))
+            .connect()
+            .await?;
+        let mut geyser = client.subscribe_once(config.request.to_proto()).await?;
 
         // Receive-send loop
         let mut send_tasks = JoinSet::new();